requirements.txt
# flask==2.1.1
flask-cors==3.0.10
flask-restx==0.5.1
pymongo==4.1.1